+++ ent/testplan/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.TestPlan {
+func ID(id int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.TestPlan {
+func IDEQ(id int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.TestPlan {
+func IDNEQ(id int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.TestPlan {
+func IDIn(ids ...int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.TestPlan {
+func IDNotIn(ids ...int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.TestPlan {
+func IDGT(id int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.TestPlan {
+func IDGTE(id int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.TestPlan {
+func IDLT(id int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.TestPlan {
+func IDLTE(id int64) predicate.TestPlan {
 	return predicate.TestPlan(sql.FieldLTE(FieldID, id))
 }
